Adding initial class and ideas for composite.
Composite cipher is required and should hold list of user selections
then execute them all in order however more 'files' are required.
This is because each will need to output to 'temp' files until the list
is completed. Also, keys will need to be considered (how to manage list
etc.)